refactor(nav): split dom-swapper into focused dom modules
Decompose the browser-router DOM swapper into html-parser, stylesheet-preloader,
head-updater, body-morpher, and a thin DomSwapper facade under client/dom/.
