+++ packages/browser-router/src/client/services/index.ts
@@ -3,6 +3,6 @@
  * @module
  */
 
-export { DomSwapper } from './dom-swapper.ts';
+export { DomSwapper } from '../dom/dom-swapper.ts';
 export { ViewTransitionManager } from './view-transition-manager.ts';
 export { PrefetchManager } from './prefetch-manager.ts';
